requirements.txt
beautifulsoup4==4.14.3
requests==2.32.5
selenium==4.41.0
python-dotenv==1.2.2
pandas==2.3.3
lxml==5.4.0